refactor(stream): re-organize TopN code
