+++ src/stream/src/executor/mod.rs
@@ -58,7 +58,6 @@ mod error;
 mod expand;
 mod filter;
 mod global_simple_agg;
-mod group_top_n;
 mod hash_agg;
 pub mod hash_join;
 mod hop_window;
@@ -77,8 +76,6 @@ mod sink;
 pub mod source;
 pub mod subtask;
 mod top_n;
-mod top_n_appendonly;
-mod top_n_executor;
 mod union;
 mod wrapper;
 
@@ -97,7 +94,6 @@ pub use error::{StreamExecutorError, StreamExecutorResult};
 pub use expand::ExpandExecutor;
 pub use filter::FilterExecutor;
 pub use global_simple_agg::GlobalSimpleAggExecutor;
-pub use group_top_n::GroupTopNExecutor;
 pub use hash_agg::HashAggExecutor;
 pub use hash_join::*;
 pub use hop_window::HopWindowExecutor;
@@ -115,8 +111,7 @@ use risingwave_pb::source::{ConnectorSplit, ConnectorSplits};
 use simple::{SimpleExecutor, SimpleExecutorWrapper};
 pub use sink::SinkExecutor;
 pub use source::*;
-pub use top_n::TopNExecutor;
-pub use top_n_appendonly::AppendOnlyTopNExecutor;
+pub use top_n::{AppendOnlyTopNExecutor, GroupTopNExecutor, TopNExecutor};
 pub use union::UnionExecutor;
 pub use wrapper::WrapperExecutor;
